.gitignore 파일 수정, models -> entities로 디렉토리 네임 변경
pycharm 프로젝트에 추가하면서 .idea 폴더 git에서 관리하지 않도록 .gitignore 파일 수정 하고 models 폴더 명 entities로 변경
